fix: adressing comments on PR
