+++ application/CohortManager/compose.cohort-distribution.yaml
@@ -12,6 +12,7 @@ services:
       - ASPNETCORE_URLS=http://*:7077
       - DtOsDatabaseConnectionString=Server=localhost,1433;Database=${DB_NAME};User Id=SA;Password=${PASSWORD};TrustServerCertificate=True
       - ExceptionFunctionURL=http://localhost:7070/api/CreateException
+      - CohortDistributionDataServiceURL=http://localhost:7915/api/CohortDistributionDataService/
 
   retrieve-cohort-distribution-data:
     container_name: retrieve-cohort-distribution-data
