+++ application/CohortManager/src/Functions/Shared/Common/Interfaces/ICreateCohortDistributionData.cs
@@ -5,7 +5,6 @@ using Model.DTO;
 
 public interface ICreateCohortDistributionData
 {
-    Task<bool> InsertCohortDistributionData(CohortDistributionParticipant cohortDistributionParticipant);
     List<CohortDistributionParticipantDto> GetUnextractedCohortDistributionParticipants(int rowCount);
     bool UpdateCohortParticipantAsInactive(string NhsNumber);
     CohortDistributionParticipant GetLastCohortDistributionParticipant(string NhsNumber);
